Rename new branch files to camelCase.
Align newly added files with the repository's camelCase file naming
convention and update the MeetingCallingView import path accordingly.
